traffic_monitoring/requirements.txt
numpy>=1.20.0
opencv-python>=4.6.0
onnx>=1.12.0
onnxruntime>=1.13.1
ultralytics>=8.0.0
boxmot>=10.0.33
easyocr>=1.7.0
paho-mqtt>=2.2.1
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
python-multipart>=0.0.6
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=7.3.1